% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulants.R
\name{cumulants}
\alias{cumulants}
\title{Moment function}
\usage{
cumulants(k, dist, param, domain = "realline", rel.tol = 1e-06)
}
\description{
Calcule theoretical raw moments, central moments, absolute moments or absolute central moments of any continuous distribution.
}
\examples{
# Enter example D: D: D:
}
\seealso{
\code{\link{Distributions}} for other standard distributions.
}
\author{
Jorge Iván Pérez, \email{jivan.perez@udea.edu.co}
}
