% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurt.R
\name{kurt}
\alias{kurt}
\title{Kurtosis function}
\usage{
kurt(excess = TRUE, dist, param, domain = "realline",
  rel.tol = 1e-06)
}
\description{
Calcule theoretical kurtosis or excess kurtosis of any continuous distribution.
}
\examples{
# Enter example D: D: D:
}
\seealso{
\code{\link{Distributions}} for other standard distributions.
}
\author{
Jorge Iván Pérez, \email{jivan.perez@udea.edu.co}
}
