% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{moments}
\alias{moments}
\title{Moment function}
\usage{
moments(k, dist, param, domain, central = FALSE, absolute = FALSE)
}
\arguments{
\item{k}{order of the moment of interest}

\item{dist}{density name for the distribution. The created density functions must have a name of the form \code{dxxx}. To understand its use see details and examples.}

\item{param}{are the parameters of the distribution. The name of each parameter must be specified. To understand its use see examples.}

\item{domain}{defines the domain of the distribution function. The type of domain of distribution to be tried see details.}

\item{central}{logical; if TRUE, the k-th central moments are given as \eqn{E[(X-\mu)^k]}.}

\item{absolute}{logical; if TRUE, the k-th absolute moments are given as \eqn{E|X|^k]}.}
}
\value{
\code{moments} gives the theorical k-th raw, central, absolute or absolute central moments of any continuous probability distribution function.
}
\description{
Calcule theoretical raw, central, absolute or absolute central moments of any continuous probability distribution function.
}
\details{
The \code{moments} function supports probability distribution functions of a large number of libraries.

In the \code{dist} argument, you must enter the name of the distribution of interest, for example, you can enter \code{"gamma"} or \code{"dgamma"}, both will produce the same result.

If \eqn{f(x)} has no parameters, then do \code{param = NULL}.

The following are the different \code{domain} argument: \itemize{
\item \code{binom}: for discrete distributions of binomial type.
\item \code{counts}: for discrete distributions of counting type.
\item \code{realline}: for continuous distributions defined between -\eqn{\infty} and \eqn{\infty}. 
\item \code{realplus}: for continuous distributions defined between 0 and \eqn{\infty}.
\item \code{real0to1}: for continuous distributions defined between 0 and 1.
\item \code{real-1to1}: for continuous distributions defined between -1 and 1.
\item \code{c(lower = a, upper = b)}: for continuous distributions defined between \code{a} and \code{b}.
}

If \code{central = TRUE} and \code{absolute = TRUE} are selected, the k-th central absolute moments is calculated and given as \eqn{E|(X-\mu)^k|}.
}
\note{
many distributions support \code{domain = "realline"} even though they are not defined from -\eqn{\infty} to \eqn{\infty} because of their programming. It is recommended to try initially with this argument.
}
\examples{
# Let's try first with distributions of the library stats
moments(k = 1:4, dist = "dchisq", param = c(df = 3), domain = "realplus")
# or
moments(k = 1:4, dist = "chisq", param = c(df = 4), domain = "realplus")

#---------------------------------------------------------------------------------------

# the name of the created density functions must have a name
# of the form dxxx. Also, how does it not have parameters
# then \\code{param = NULL}
dmyfunction <- function(x) x^3/4 
# so that it integrates to 1, x must be between 0 to 2.
moments(k = 1:4, dist = "dmyfunction", param = NULL, domain = c(0, 2))

#---------------------------------------------------------------------------------------

# Let's try distributions from other libraries
if(!require("extraDistr")) install.packages("extraDistr") # to install the package
# The same result is obtained with the diferent domain (see 'Note')
moments(k = 1:2, dist = "dpareto", param = c(a = 3, b = 7),
        domain = "realline")
# or
moments(k = 1:2, dist = "dpareto", param = c(a = 3, b = 7),
        domain = c(7, Inf))
# In this case, no moments are calculated for k> 2, because the
# parameter of the pareto distribution is \\code {a = 3}, and
# therefore, the moments are defined for \\eqn{E (X ^ k) < a}.
# Read about pareto distribution for more information.

#---------------------------------------------------------------------------------------

# Let's try distributions from other libraries to calculated central
# and absolute moments
if(!require("gamlss.dist")) install.packages("gamlss.dist") # to install the package
moments(k = 3, dist = "PE", param = c(mu = -25, sigma = 7, nu = 4),
        central = TRUE) 
moments(k = 3, dist = "PE", param = c(mu = -25, sigma = 7, nu = 4),
        absolute = TRUE)
moments(k = 3, dist = "PE", param = c(mu = -25, sigma = 7, nu = 4),
        central = TRUE, absolute = TRUE)

#---------------------------------------------------------------------------------------

# Let's try with a discrete counting distribution to calculated
# raw and central moments
if(!require("gamlss.dist")) install.packages("gamlss.dist") # to install the package
moments(k = 1:4, dist = "DEL", param = c(mu = 2, sigma = 3, nu = 0.5),
        domain = "counts")
moments(k = 1:4, dist = "DEL", param = c(mu = 2, sigma = 3, nu = 0.5),
        domain = "counts", central = TRUE)
        
#---------------------------------------------------------------------------------------

# Let's try with a discrete binomial type distribution to calculated
# raw and central moments
moments(k = 1:4, dist = "binom", param = c(size = 15, prob = 0.3),
        domain = "binom")
moments(k = 1:4, dist = "binom", param = c(size = 15, prob = 0.3),
        domain = "binom", central = TRUE)
}
\seealso{
\code{\link{Distributions}} for other standard distributions.

\code{\link{cumulants}}, \code{\link{skew}}, \code{\link{kurt}}.
}
\author{
Jorge Iván Pérez, \email{jivan.perez@udea.edu.co}
}
